+++ src/Symfony/Component/HttpKernel/CHANGELOG.md
@@ -18,6 +18,7 @@ CHANGELOG
  * Add the `$excludedPaths` and `$excludedHttpCodes` arguments to `ProfilerListener::__construct()` to skip profiling some requests
  * Make `#[MapRequestPayload]` deserialize any media type carrying a structured syntax suffix with the encoder of the suffix format, e.g. `application/vnd.api+json` with the `json` encoder
  * Add the `$exposeHeaders` argument to `#[RateLimit]`
+ * Instantiate on demand the bundles that have nothing to do when the kernel boots
 
 8.1
 ---